+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2026-01-06
+
+### Added
+- Store URL is now displayed in PR comment notifications alongside Theme ID, Preview URL, and Admin URL
+
 ## [1.0.0] - 2025-01-01
 
 ### Added